+++ src/lib/components/StatusBar.svelte
@@ -65,10 +65,6 @@
 
 <style>
   .statusbar {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
     display: flex;
     justify-content: space-between;
     align-items: center;